// codeapp/src/generated/services/Pvci_transcriptsessionsService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import type { Pvci_transcriptsessionsBase, Pvci_transcriptsessions } from '../models/Pvci_transcriptsessionsModel';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Pvci_transcriptsessionsService {
  private static readonly dataSourceName = 'pvci_transcriptsessions';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Pvci_transcriptsessionsBase, 'pvci_transcriptsessionid'>): Promise<IOperationResult<Pvci_transcriptsessions>> {
    const result = await Pvci_transcriptsessionsService.client.createRecordAsync<Omit<Pvci_transcriptsessionsBase, 'pvci_transcriptsessionid'>, Pvci_transcriptsessions>(
      Pvci_transcriptsessionsService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Pvci_transcriptsessionsBase, 'pvci_transcriptsessionid'>>): Promise<IOperationResult<Pvci_transcriptsessions>> {
    const result = await Pvci_transcriptsessionsService.client.updateRecordAsync<Partial<Omit<Pvci_transcriptsessionsBase, 'pvci_transcriptsessionid'>>, Pvci_transcriptsessions>(
      Pvci_transcriptsessionsService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Pvci_transcriptsessionsService.client.deleteRecordAsync(
      Pvci_transcriptsessionsService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Pvci_transcriptsessions>> {
    const result = await Pvci_transcriptsessionsService.client.retrieveRecordAsync<Pvci_transcriptsessions>(
      Pvci_transcriptsessionsService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Pvci_transcriptsessions[]>> {
    const result = await Pvci_transcriptsessionsService.client.retrieveMultipleRecordsAsync<Pvci_transcriptsessions>(
      Pvci_transcriptsessionsService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Pvci_transcriptsessions> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Pvci_transcriptsessionsService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Pvci_transcriptsessionsService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}