// codeapp/src/generated/models/Pvci_agentinventoriesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pvci_agentinventoriesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pvci_agentinventoriesstatecode = keyof typeof Pvci_agentinventoriesstatecode;
export const Pvci_agentinventoriesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pvci_agentinventoriesstatuscode = keyof typeof Pvci_agentinventoriesstatuscode;

export interface Pvci_agentinventoriesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  pvci_agentinventoryid: string;
  pvci_authoringorigin?: string;
  pvci_botid?: string;
  pvci_classificationconfidence?: string;
  pvci_classificationsource?: string;
  pvci_displayname?: string;
  pvci_environmentid?: string;
  pvci_environmentname?: string;
  pvci_environmenturl?: string;
  pvci_evidencejson?: string;
  pvci_harness?: string;
  pvci_inventorysource?: string;
  pvci_lastsyncedon?: string;
  pvci_model?: string;
  pvci_name: string;
  pvci_orchestrationtype?: string;
  pvci_published?: boolean;
  pvci_resourceid?: string;
  pvci_resourcetype?: string;
  pvci_schemaname?: string;
  pvci_sourcekey?: string;
  pvci_tenantid?: string;
  statecode: Pvci_agentinventoriesstatecode;
  statuscode?: Pvci_agentinventoriesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pvci_agentinventories extends Pvci_agentinventoriesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  pvci_publishedname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}