// codeapp/src/generated/models/Pvci_transcriptsessionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pvci_transcriptsessionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pvci_transcriptsessionsstatecode = keyof typeof Pvci_transcriptsessionsstatecode;
export const Pvci_transcriptsessionsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pvci_transcriptsessionsstatuscode = keyof typeof Pvci_transcriptsessionsstatuscode;

export interface Pvci_transcriptsessionsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pvci_activitiesjson?: string;
  pvci_activitycount?: number;
  pvci_agentturncount?: number;
  pvci_botid?: string;
  pvci_botname?: string;
  pvci_channel?: string;
  pvci_comments?: string;
  pvci_conversationjson?: string;
  pvci_correlationstatus?: string;
  pvci_csat?: string;
  pvci_datasource?: string;
  pvci_durationseconds?: number;
  pvci_embeddedconversationguid?: string;
  pvci_enddatetimeutc?: string;
  pvci_eventcount?: number;
  pvci_ingestedon?: string;
  pvci_initialusermessage?: string;
  pvci_isresolvedimplied?: string;
  pvci_istestmode?: boolean;
  pvci_lastagentmessage?: string;
  pvci_messagecount?: number;
  pvci_metadatajson?: string;
  pvci_monitorsessionid?: string;
  pvci_multiuseranomaly?: boolean;
  pvci_name: string;
  pvci_outcomereason?: string;
  pvci_parsedturnsjson?: string;
  pvci_payloadtruncated?: boolean;
  pvci_planeventsjson?: string;
  pvci_rawchattranscript?: string;
  pvci_sessionoutcome?: string;
  pvci_startdatetimeutc?: string;
  pvci_tenantid?: string;
  pvci_topicid?: string;
  pvci_topicname?: string;
  pvci_transcriptcreatedon?: string;
  pvci_transcriptid?: string;
  pvci_transcriptsessionid: string;
  pvci_turncount?: number;
  pvci_useraadobjectid?: string;
  pvci_userdisplayname?: string;
  "pvci_UserId@odata.bind"?: string;
  pvci_userturncount?: number;
  pvci_userupn?: string;
  statecode: Pvci_transcriptsessionsstatecode;
  statuscode?: Pvci_transcriptsessionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pvci_transcriptsessions extends Pvci_transcriptsessionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pvci_istestmodename?: string;
  pvci_multiuseranomalyname?: string;
  pvci_payloadtruncatedname?: string;
  pvci_useridname?: string;
  pvci_useridyominame?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pvci_userid?: object;
  _pvci_userid_value?: string;
}